feat(feature/donkey/#59): 게시물 정보 최상위 컴포넌트로 이동, 헤더에 게시물 수 표시
