+++ instagram/src/components/UserPage/PostsFollows.jsx
@@ -4,11 +4,11 @@ import {
   ListNumber,
 } from "../../styles/UserPage/PostsFollowsStyle";
 
-function PostsFollows() {
+function PostsFollows({ postData }) {
   return (
     <ListBox>
       <List>
-        게시물<ListNumber>{0}</ListNumber>
+        게시물<ListNumber>{`${Object.keys(postData).length}`}</ListNumber>
       </List>
       <List>
         팔로워<ListNumber>{0}</ListNumber>
