+++ instagram/src/components/UserPage/UserHeader.jsx
@@ -2,13 +2,10 @@ import ProfilePic from "./ProfilePic";
 import HeaderSection from "./HeaderSection";
 import { HeaderContainer } from "../../styles/UserPage/UserHeaderStyle";
 import { useEffect, useState } from "react";
-import { getUserNo } from "./../../utils/getToken";
 import axios from "../../api/config";
 
-function UserHeader() {
+function UserHeader({ userNo, postData }) {
   const [userInfo, setUserInfo] = useState({});
-  const userNo = getUserNo();
-
   useEffect(() => {
     axios.get(`user/profile/${userNo}`).then((res) => {
       setUserInfo({ ...res.data.userInfo });
@@ -23,6 +20,7 @@ function UserHeader() {
           <HeaderSection
             userName={userInfo.name}
             userNickname={userInfo.nickname}
+            postData={postData}
           />
         </HeaderContainer>
       ) : null}
